refactor: domain 폴더 옮기기
