+++ src/main/java/com/openhack/toyland/domain/skill/Skill.java
@@ -1,4 +1,4 @@
-package com.openhack.toyland.skill;
+package com.openhack.toyland.domain.skill;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,7 +6,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 
-import com.openhack.toyland.BaseTimeEntity;
+import com.openhack.toyland.domain.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
