+++ src/main/java/com/openhack/toyland/domain/skill/TechStack.java
@@ -1,4 +1,4 @@
-package com.openhack.toyland.user;
+package com.openhack.toyland.domain.skill;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,7 +6,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 
-import com.openhack.toyland.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,14 +13,14 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class User extends BaseTimeEntity {
+public class TechStack {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private Long githubIdentifier;
+    private Long toyId;
 
     @NotNull
-    private String username;
+    private Long skillId;
 }
