+++ src/main/java/com/openhack/toyland/domain/skill/TechStackRepository.java
@@ -1,4 +1,4 @@
-package com.openhack.toyland.skill;
+package com.openhack.toyland.domain.skill;
 
 import org.springframework.data.jpa.repository.JpaRepository;
