+++ src/main/java/com/openhack/toyland/domain/toy/Category.java
@@ -1,0 +1,11 @@
+package com.openhack.toyland.domain.toy;
+
+public enum Category {
+	WEB,
+	APP,
+	IOS,
+	ANDROID,
+	AI,
+	ML,
+	BIG_DATA;
+}
