+++ src/main/java/com/openhack/toyland/domain/toy/ToyRepository.java
@@ -1,4 +1,4 @@
-package com.openhack.toyland.toy;
+package com.openhack.toyland.domain.toy;
 
 import org.springframework.data.jpa.repository.JpaRepository;
