+++ src/main/java/com/openhack/toyland/domain/user/User.java
@@ -1,6 +1,4 @@
-package com.openhack.toyland;
-
-import java.time.LocalDateTime;
+package com.openhack.toyland.domain.user;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -8,6 +6,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
 
+import com.openhack.toyland.domain.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,15 +14,14 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Maintenance extends BaseTimeEntity {
+public class User extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private Long toyId;
+    private Long githubIdentifier;
 
-    private Boolean healthCheck;
-
-    private LocalDateTime active;
+    @NotNull
+    private String username;
 }
