+++ src/main/java/com/openhack/toyland/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.openhack.toyland.user;
+package com.openhack.toyland.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
