--- src/main/java/com/openhack/toyland/toy/Category.java
@@ -1,11 +1,0 @@
-package com.openhack.toyland.toy;
-
-public enum Category {
-    WEB,
-    APP,
-    IOS,
-    ANDROID,
-    AI,
-    ML,
-    BIG_DATA
-}
